SDXL/Content/Python/requirements.txt
accelerate==0.30.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
coloredlogs==15.0.1
compel==2.0.2
cycler==0.12.1
diffusers==0.27.0
filelock==3.13.1
flatbuffers==24.3.25
fsspec==2024.3.0
huggingface-hub==0.21.4
humanfriendly==10.0
idna==3.6
importlib_metadata==7.0.2
intel-openmp==2021.4.0
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
mkl==2021.4.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
onnxruntime==1.17.3
opencv-python==4.9.0.80
pillow==10.2.0
protobuf==5.26.1
psutil==5.9.8
pyparsing==3.1.2
pyreadline3==3.4.1
PySide6==6.6.2
PySide6_Addons==6.6.2
PySide6_Essentials==6.6.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
shiboken6==6.6.2
sympy==1.12
tbb==2021.12.0
tokenizers==0.15.2
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
tqdm==4.66.2
transformers==4.38.2
typing_extensions==4.10.0
urllib3==2.2.1
zipp==3.18.1
